HH-218702 move DelegatingDataSource to nab-jdbc-common module
