+++ nab-data-source/src/main/java/ru/hh/nab/datasource/NamedDataSource.java
@@ -3,6 +3,7 @@ package ru.hh.nab.datasource;
 import java.sql.SQLException;
 import java.util.Optional;
 import javax.sql.DataSource;
+import ru.hh.nab.jdbc.common.datasource.DelegatingDataSource;
 
 /**
  * Wrapper for DataSource which holds unique {@link #getName() name}.
