+++ nab-data-source/src/main/java/ru/hh/nab/datasource/monitoring/StatementTimeoutDataSource.java
@@ -4,7 +4,7 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
 import javax.sql.DataSource;
-import ru.hh.nab.datasource.DelegatingDataSource;
+import ru.hh.nab.jdbc.common.datasource.DelegatingDataSource;
 
 /**
  * A {@link javax.sql.DataSource} that sends Postgresql statement timeout query after obtaining a connection from the underlying datasource.<br/>
